feat: frontend - refactor better practices
