+++ src/shared/AlertMessage/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Alert } from '@edx/paragon';
+
 import './index.scss';
 
 const AlertMessage = ({
