+++ src/shared/LabDetailsCard/index.scss
