+++ src/shared/LabDetailsChartCard/index.scss
