+++ src/views/LabDetails/index.scss
