+++ src/views/LabSummary/index.scss
